lapa/frontends/swift.py
"""
Swift Frontend for LAPA.

This module provides the frontend implementation for the Swift programming language.
"""

from lapa.frontend import Frontend

class SwiftFrontend(Frontend):
    """Frontend for the Swift programming language."""

    def __init__(self):
        """Initialize the Swift frontend."""
        super().__init__()
        self.language = "Swift"
        self.file_extensions = {".swift"}

    def parse(self, source_code: str):
        """
        Parse Swift source code and return an abstract syntax tree (AST).

        Args:
            source_code: The Swift source code to parse.

        Returns:
            An abstract syntax tree (AST) representation of the source code.
        """
        # TODO: Integrate Swift compiler or parser to generate AST
        raise NotImplementedError("Swift parsing is not yet implemented.")

    def ast_to_ir(self, ast):
        """
        Convert a Swift AST to the intermediate representation (IR).

        Args:
            ast: The abstract syntax tree to convert.

        Returns:
            The intermediate representation (IR) of the AST.
        """
        # TODO: Implement AST to IR conversion for Swift
        raise NotImplementedError("Swift AST to IR conversion is not yet implemented.")

    def parse_file(self, file_path: str):
        """
        Parse a Swift file and return an abstract syntax tree (AST).

        Args:
            file_path: The path to the Swift file to parse.

        Returns:
            An abstract syntax tree (AST) representation of the file.
        """
        # TODO: Implement file parsing for Swift
        raise NotImplementedError("Swift file parsing is not yet implemented.")

    def parse_string(self, source_code: str):
        """
        Parse Swift source code from a string and return an abstract syntax tree (AST).

        Args:
            source_code: The Swift source code to parse.

        Returns:
            An abstract syntax tree (AST) representation of the source code.
        """
        # TODO: Implement string parsing for Swift
        raise NotImplementedError("Swift string parsing is not yet implemented.")
